Transfer from Private Repository
Enclosed is a project based on enviromental data of the West Coast of
Florida.  The goal was to generate an animation that tracks enviromental
factors that could contribute to respiratory issues on the beach.
Respiratory distress is measured by trained life guards.  The enclosed
animation is a pass at investigating the relationship between sea
conditions and possible respiratory issues.